Fix phpinfo setup and PHP module enable

+++ LAMP-stack.sh
@@ -15,10 +15,10 @@ echo "
 "
 sleep 0.5
 echo "                                                        
-Author: 	rock-dotcom
+Author: 	Kimi Autto (github.com/Z-A-P-P-I-T)
 Creation Date:  21/3-21
-Description: 	Script that will install a Lampstack server 
-		with phpmyadmin for Ubuntu and Debian Distros 
+Description: 	Script that will install a LAMP stack server 
+		with phpMyAdmin for Debian and Ubuntu distros 
 "
 
 echo "
@@ -26,7 +26,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |     This Script will Install a LAMP Stack Server    |       
-       |                  On Ubuntu 20.04                    |
+       |                 On Debian/Ubuntu                    |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -42,7 +42,7 @@ echo "
        +------------------.oooO--(_)--Oooo.------------------+
        |      Lest start By Running Update and Upgrade       |
        |            Please provide your Password             |       
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -66,7 +66,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |              Update and Upgrade Done                |
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -81,7 +81,7 @@ echo "
        +------------------.oooO--(_)--Oooo.------------------+
        |     Now we open ports 22 (for SSH), 80 and 443      |
        |          and enable Ubuntu Firewall (ufw)           |       
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -102,7 +102,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |   Ports have been opened and the Firewall enabled   |
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -117,7 +117,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |       Installing and Testing Apache Webserver       |
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -125,7 +125,11 @@ echo "
                                     (_/
 "
 sleep 2
-sudo systemctl stop apache2
+if systemctl list-unit-files --type=service | awk '{print $1}' | grep -qx "apache2.service"; then
+    sudo systemctl stop apache2
+else
+    echo "Apache service not loaded yet; skipping stop."
+fi
 sleep 4
 sudo apt install -y apache2
  
@@ -138,8 +142,8 @@ echo "
                                 ''~``
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
-       |           Installing and testing PHP 7.4            |
-       |                    Ubuntu 20.04                     |
+       |           Installing and testing PHP                |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -148,7 +152,7 @@ echo "
 "
 sleep 2
 sudo apt install -y php php-mysql php-cli php-json php-common php-opcache libapache2-mod-php
- 
+
 sleep 2
 clear
 echo "
@@ -156,7 +160,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |           Checking Installed PHP Version.           |
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -165,7 +169,22 @@ echo "
 "
 php --version
 echo '<?php phpinfo(); ?>' | sudo tee /var/www/html/phpinfo.php > /dev/null
- 
+sudo chmod 644 /var/www/html/phpinfo.php
+
+# Ensure Apache loads PHP module
+php_version=$(php -r 'echo PHP_MAJOR_VERSION . "." . PHP_MINOR_VERSION;' 2>/dev/null || echo "")
+if [[ -n "$php_version" ]]; then
+    php_module="php${php_version}"
+    if [[ -f "/etc/apache2/mods-available/${php_module}.load" ]]; then
+        sudo a2enmod "$php_module" >/dev/null 2>&1 || true
+    else
+        echo "PHP Apache module not found (${php_module})."
+        echo "If phpinfo.php fails, install libapache2-mod-php and re-run."
+    fi
+fi
+sudo a2dismod mpm_event >/dev/null 2>&1 || true
+sudo a2enmod mpm_prefork >/dev/null 2>&1 || true
+
 sleep 2
 clear
  
@@ -174,7 +193,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |           Restarting Apache Please Wait.            |       
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -206,7 +225,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |                 Installing MariaDB                  |       
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -223,7 +242,7 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |         Checking Installed MariaDB Version.         |       
-       |                    Ubuntu 20.04                     |
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -240,8 +259,8 @@ echo "
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
        |             Securing MariaDB Installation           |
-       |  Please fallow the instructions and set a Password  |       
-       |                    Ubuntu 20.04                     |
+       |  Please follow the instructions and set a Password  |       
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -257,8 +276,8 @@ echo "
                                 ''~``
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
-       |         Installing and Verifiying phpMyAdmin        |      
-       |                    Ubuntu 20.04                     |
+       |         Installing and Verifying phpMyAdmin         |      
+       |                 Debian/Ubuntu                       |
        |                    .oooO                            |
        |                    (   )   Oooo.                    |
        +---------------------\ (----(   )--------------------+
@@ -274,7 +293,7 @@ echo "
                                 ''~``
                                ( o o )
        +------------------.oooO--(_)--Oooo.------------------+
-       |                   Congratulation!                   |
+       |                   Congratulations!                  |
        |   You should now have a LAMP stack webserver with   |
        |  Linux, Apache, MariaDB and PHP set up and running! | 
        |  You are now ready to start building your website.  |
